refactor: move list_apps.py's helpers onto Site
list_installed_apps/query_installed_apps_via_db/_query_installed_apps_via_frappe
were already CLI-free but lived under pilot/commands — moved to
pilot/core/site.py with a new Site.installed_apps() wrapper.
query_installed_apps_via_db is now public (was the underscore-prefixed
_query_via_db_cli that admin/backend/providers/sites.py was already
reaching past). ListSiteAppsCommand and SiteProvider both call the new
home instead of pilot.commands.
